pyproject.toml
[project]
name = "founder-matching"
version = "0.1.0"
description = "Real-time cofounder matchmaking system for structured networking events"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "redis[hiredis]>=5.2",
    "networkx>=3.4",
    "anthropic>=0.42",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.18",
    "httpx>=0.28",
    "qrcode[pil]>=8.0",
    "reportlab>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "fakeredis>=2.25",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
